Day04/spring-controller/src/test/java/com/swarangi/app/HelloControllerTest.java
package com.swarangi.app;

import org.junit.Assert;
import org.junit.Test;




public class HelloControllerTest {
    HelloController helloController=new HelloController();
    @Test

    public void testSayHello(){
        String expectedResult = "Hello from App";
        String result=helloController.SayHello();
        Assert.assertEquals(expectedResult,result);
    }

    @Test

    public void testrequestHello(){
        String expectedResult = "Hello from Request mapping";
        String result=helloController.requestHello();
        Assert.assertEquals(expectedResult,result);
    }
    @Test
    public void testrequestWelcome1(){
        String name="world";
        String expectedResult = "welcome to "+name;;
        String result= helloController.welcome(name);
        Assert.assertEquals(expectedResult,result);
    }
    @Test
    public void testrequestSendMessage(){
        String message="world";
        String expectedResult = "send message"+message;
        String result= helloController.sendMessage(message);
        Assert.assertEquals(expectedResult,result);
    }


}
